+++ src/app/[locale]/(website)/blog/[slug]/page.tsx
@@ -1,19 +1,19 @@
 "use client";
 
-import { Footer } from "@/src/app/components/footer";
-import { Paragraph } from "@/src/app/components/paragraph";
+import { Footer } from "@/src/app/[locale]/components/footer";
+import { Paragraph } from "@/src/app/[locale]/components/paragraph";
 import imageLoader from "@/src/app/utils/image-loader";
 import Image from "next/image";
-import { Heading } from "@/src/app/components/heading";
-import { PageWrapper } from "@/src/app/components/page-wrapper";
+import { Heading } from "@/src/app/[locale]/components/heading";
+import { PageWrapper } from "@/src/app/[locale]/components/page-wrapper";
 import { Post } from "@/src/app/utils/interface";
 import { PortableText, PortableTextComponents } from "next-sanity";
 import React, { useMemo } from "react";
 import { formatDate } from "@/src/app/utils/date";
 import { useParams } from "next/navigation";
 import { useSanityData } from "@/src/app/utils/use-sanity-data";
-import { LoadingAnimation } from "@/src/app/components/loading-animation";
-import { InlineError } from "@/src/app/components/inline-error";
+import { LoadingAnimation } from "@/src/app/[locale]/components/loading-animation";
+import { InlineError } from "@/src/app/[locale]/components/inline-error";
 
 const GET_BLOG_POST = `*[_type == "post" && slug.current == $slug][0] {
   title,
