+++ src/app/[locale]/(website)/blog/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Footer } from "@/src/app/components/footer";
+import { Footer } from "@/src/app/[locale]/components/footer";
 import Blogs from "./blogs";
 
 export default function Page() {
