+++ src/app/[locale]/(website)/dogodki/page.tsx
@@ -1,4 +1,4 @@
-import { Footer } from "@/src/app/components/footer";
+import { Footer } from "@/src/app/[locale]/components/footer";
 import Hero from "./hero";
 import { NextEvent } from "../next-event";
 import PastEvents from "./past-events";
