+++ src/app/[locale]/(website)/layout.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
-import { Navbar } from "@/src/app/components/navbar";
-import { LoadingAnimation } from "@/src/app/components/loading-animation";
+import { Navbar } from "@/src/app/[locale]/components/navbar";
+import { LoadingAnimation } from "@/src/app/[locale]/components/loading-animation";
 
 export default function Layout({
   children,
