+++ src/app/[locale]/(website)/loading.tsx
@@ -1,5 +1,5 @@
-import { LoadingAnimation } from "@/src/app/components/loading-animation";
-import { PageWrapper } from "@/src/app/components/page-wrapper";
+import { LoadingAnimation } from "@/src/app/[locale]/components/loading-animation";
+import { PageWrapper } from "@/src/app/[locale]/components/page-wrapper";
 
 export default function Loading() {
   return (
