+++ src/app/[locale]/(website)/partnerstvo/partners-list.tsx
@@ -1,5 +1,5 @@
-import { PageWrapper } from "@/src/app/components/page-wrapper";
-import { Paragraph } from "@/src/app/components/paragraph";
+import { PageWrapper } from "@/src/app/[locale]/components/page-wrapper";
+import { Paragraph } from "@/src/app/[locale]/components/paragraph";
 import { useTranslations } from "next-intl";
 import Image from "next/image";
