+++ src/app/[locale]/(website)/partnerstvo/statistics.tsx
@@ -1,6 +1,6 @@
-import { Heading } from "@/src/app/components/heading";
-import { PageWrapper } from "@/src/app/components/page-wrapper";
-import { StatsCard } from "@/src/app/components/stats-card";
+import { Heading } from "@/src/app/[locale]/components/heading";
+import { PageWrapper } from "@/src/app/[locale]/components/page-wrapper";
+import { StatsCard } from "@/src/app/[locale]/components/stats-card";
 import { useTranslations } from "next-intl";
 import Image from "next/image";
