+++ src/app/[locale]/components/activity-card.tsx
