+++ src/app/[locale]/components/blog-card.tsx
