+++ src/app/[locale]/components/card.tsx
