+++ src/app/[locale]/components/footer-newsletter.tsx
