+++ src/app/[locale]/components/footer.tsx
