+++ src/app/[locale]/components/heading.tsx
