+++ src/app/[locale]/components/loading-animation.tsx
