+++ src/app/[locale]/components/navbar-menu.tsx
