+++ src/app/[locale]/components/navbar.tsx
