+++ src/app/[locale]/components/newsletter-component-2.tsx
