+++ src/app/[locale]/components/newsletter-component.tsx
