+++ src/app/[locale]/components/paragraph.tsx
