+++ src/app/[locale]/components/post-component.tsx
