+++ src/app/[locale]/components/review-component.tsx
