+++ src/app/[locale]/components/skeleton.tsx
