+++ src/app/[locale]/components/stats-card.tsx
