+++ src/app/[locale]/components/team-member-component.tsx
